Add UI

+++ Login/Common/DTO/ProductDTO.cs
@@ -11,11 +11,11 @@ namespace Login.Common.DTO
         public string Name { get; set; }
         public long DiscountId { get; set; }
         public string Barcode { get; set; }
-        public int AmountInPackage { get; set; }
-        public int Amount { get; set; }
-        public int ActualPrice { get; set; }
-        public int Price { get; set; }
-        public int PriceOfPiece { get; set; }
+        public decimal AmountInPackage { get; set; }
+        public decimal Amount { get; set; }
+        public decimal ActualPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal PriceOfPiece { get; set; }
         public bool Selected { get; set; }
     }
 }
