+++ Login/Data/Models/Product.cs
@@ -10,21 +10,13 @@ namespace Login.Data.Models
     public class Product :BaseEntity
     {
         public string Name { get; set; }
-
-        // public string MxikCode { get; set; }
-        //public int GroupId { get; set; }
-        //public int MeasuredId { get; set; }
         public string Barcode { get; set; }
-        // public string VendorCode { get; set; }
-        //public string Tin { get; set; }
-        //public int TaxType { get; set; }
-        public int AmountInPackage { get; set; }
-        //public bool Marker { get; set; }
-        //public string Sku { get; set; }
-        public int Amount { get; set; }
-        public int ActualPrice { get; set; }
-        public int Price { get; set; }
-        public int PriceOfPiece { get; set; }
+        public decimal AmountInPackage { get; set; }
+        
+        public decimal Amount { get; set; }
+        public decimal ActualPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal PriceOfPiece { get; set; }
         public bool Selected { get; set; }
 
         public long? CompanyId { get; set; }
@@ -36,5 +28,13 @@ namespace Login.Data.Models
         public long? CategoryId { get; set; }
         public virtual ProductCategory? Category { get; set; }
 
+        // public string MxikCode { get; set; }
+        //public int GroupId { get; set; }
+        //public int MeasuredId { get; set; }
+        // public string VendorCode { get; set; }
+        //public string Tin { get; set; }
+        //public int TaxType { get; set; }
+        //public bool Marker { get; set; }
+        //public string Sku { get; set; }
     }
 }
